+++ Editor/IdProviders/IdProviderOptions.cs
@@ -30,7 +30,7 @@ namespace Unity.Services.Authentication.Editor
         /// <summary>
         /// The OpenID Connect Id provider configuration struct
         /// </summary>
-        public OpenIDConfig OidcConfig { get; set; }
+        internal OpenIDConfig OidcConfig { get; set; }
 
         /// <summary>
         /// Whether the client secret is needed in the target ID provider.