+++ Editor/Network/Requests/OpenIDConfig.cs
@@ -3,8 +3,14 @@ using UnityEngine;
 
 namespace Unity.Services.Authentication.Editor
 {
-    public struct OpenIDConfig
+    /// <summary>
+    /// OpenID connect Id provider configuration.
+    /// </summary>
+    struct OpenIDConfig
     {
+        /// <summary>
+        /// Issuer URL that uniquely identifies an OpenID Connect Id provider.
+        /// </summary>
         [JsonProperty("issuer")]
         public string Issuer { get; set; }
     }