+++ Editor/Utilities/IdProviderNames.cs
@@ -2,11 +2,13 @@ namespace Unity.Services.Authentication.Editor
 {
     static class IdProviderNames
     {
+        // Please keep them in alphabetical order
         internal const string Facebook = "Facebook";
         internal const string Google = "Google";
         internal const string GooglePlayGames = "Google Play Games";
         internal const string OpenIDConnect = "OpenID Connect";
         internal const string SignInWithApple = "Sign-in with Apple";
         internal const string Steam = "Steam";
+        internal const string Oculus = "Oculus";
     }
 }